Update hangman.py
Hang man game: A brief explanation for how this works: 1- the computer will choose a word at random from the word list. 2- the program will exclude words combined with dashes, -. 3- define it's length and display it to the player 'user'. 4- replace the word's letters with dashes and displays it to the user 5- the play will have 6 lives, meaning that after 6 wrong guesses the program will terminate, display the word and the player loses. 6- the player's guesses are either letters or words 7- if the player entered an invalid guess he/she will be notified and no lives will be deducted 8- the correct guesses of the word will be shown in their places instead of the dashes to make it easier for the player to know what letters he/she guessed and their positions in the word. 9- both correct guesses and wrong ones will be stored in lists, so the player can keep track of which letters he/she has chosen 10- all player's letters will be changed to lower case letters to make it readible by the program 11- if the player guessed all the letters right he/she will win.